Inline classifier::classifier submodule
